Sprint 1, part stretch:   - the day 1 main file was renamed;   - handle functions were moved in a new file.
